Test: Our test framework
Run it with `./Test`, obviously.

We add a `.py` file with a do-nothing test so that Pytest doesn't
exit with complaints and an error code.

This has been checked on Linux, and is probably fine on MacOS, too. I
am hopeful that it will work under Windows as well (when run from the
Git Bash shell that comes with Git for Windows) but we'll see.