README: Add "comments section"
It's actually just the issues (which includes PRs) board on GitHub.
That's how clever I am. :-)

It will be interesting to see how this works out (assuming anybody
even bothers to comment). One thing that occurs to me is that
“comments” suggesting that something is wrong and should be changed
can actually show the exact lines and suggest the change.

Another possibility for something like this would be to simply add
comments directly to the post via PRs. That would be basically be
pre-moderation on everything, though, and who wants to do that much
work?

I've also set the GitHub Wiki for this project to be publically
editable, in case anybody gets any interesting ideas for doing stuff
there related to this post.

Now that I think about it, perhaps it makes more sense to make a
single repo the entire blog, and have posts be files in the blog,
linked from the README. One could then start a single comment thread
(issue) for each post. Whether, by the time you add a bit of tooling,
this makes more or less sense than a standard static site generator
system (with perhaps Disqus hooked up for comments) is an open
question. It would be interesting to see how spam control would work
out with such a system.