Remove "no downstream rights restrictions" from CC-BY license summary
It was confusing because it sounds a bit like a CC-BY-SA or GPL
restriction, though I'm sure it's not meant to be. It's probably
more to prevent people adding DRM or whatever[1], but I'm just
removing this from the summary to avoid confusion.

Thanks to Stepehen J. Turnbull for the suggestion about this.

[1]: https://law.stackexchange.com/q/25667